backend/pyproject.toml
[project]
name = "eu4_modding_tools"
version = "0.5"
description = "Various scripts for modding Europa Universalis 4."

readme = "README.md"
authors = [{name = "Björn Gunnar Bryggman", email = "bjorn.bryggman@gmail.com"}]
maintainers = [{name = "Björn Gunnar Bryggman", email = "bjorn.bryggman@gmail.com"}]

classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3.12",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License"
]

requires-python = ">= 3.12"
dependencies = [
    "diffusers >= 0.29",
    "huggingface-hub>=0.24.5",
    "instructor>=1.3.7",
    "openai >= 1.35.0",
    "python-dotenv >= 1.0.1",
    "pyyaml >= 6.0.0",
    "replicate >= 0.28",
    "requests >= 2.32.3",
    "rich >= 13.7.1",
    "structlog >= 24.2.0",
    "sqlmodel >= 0.0.19",
    "tenacity >= 8.4.1",
    "toml >= 0.10",
    "tqdm >= 4.66",
    "Wand >= 0.6.13",
]

[project.urls]
Repository = "https://github.com/bjornbryggman/EU4-Modding-Tools.git"

[tool.rye]
managed = true
generate-hashes = true
dev-dependencies = [
    "coverage",
    "hypothesis",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.rye.scripts]
"check:ruff" = "ruff ."
"fix:ruff" = "ruff --fix ."
"format:ruff" = "ruff format"

[tool.pytest.ini_options]
testpaths = ["*tests*"]
addopts = "--tb=short"
xfail_strict = true
asyncio_mode = "auto"
filterwarnings = ["error"]
